requirements.txt
langchain==0.3.26
langchain-community==0.3.27
faiss-cpu
tiktoken
openai
python-dotenv
